chore: update favicon and apple-icon
feat(svg): add new svg file

This commit adds a new SVG file to the project. The SVG file
contains an image, but further details about its purpose or
content are not available from the diff.

feat(app): update icon0.svg to a new version

The icon has been updated with a new design. The new icon is a more modern and visually appealing design.

feat: update site metadata, icons and resume

The title and description in metadata was updated.
Favicons and resume were updated.